Moved services into their own package. Re-worked services to add access modifiers and for consistency.
